Content Query WebPart : Fixed WebUrls not showing (#296)
* Added the files to the new repository

* Modified the .gitignore

* Added the generated sppkg file

* Added animated gifs for the read me.

* Added another animated gif

* Added a first temporary version of the readme.md

* Update README.md

* Update README.md

* Updated the readme.md file by removing spaces between brackets in the handlebars templates examples.

* Updated the WebPart with inline template editing

Added inline template editing within the toolpane, automatic template
generation and fixed a bug related to the "Me" checkbox while filtering
user fields.

* Removed the debug folder from sources

* Updated the toolpart animated gif

* Removed toolpart animated gif

* Updated the README.md file with the latest changes.

* Update README.md

* Update README.md

* Update README.md

* Adding AceEditor as Code Editor

* Removing old Monato Artefacts

* Finalized the code editor update

Added documentation for the code editor, fixed the npm-shrinkwrap to
support the code editor dependencies, updated images for the new
documentation, regenerated the .sppkg file and updated the sources on
the public CDN for version 1.0.1

* Added a .gif for the new code editor documentation

* Updated the WebPart with the latest spfx packages.

Multiple bottle-neck bugs have been fixed in SPFX which allows to update
the WebPart with the latest SPFX packages. Overall performance should
now be better and other minor bugs have been fixed.

* ...

* Updated the readme

* Updated the readme with the solution author

* Added the External Scripts property to the toolpart

The new External Scripts property allows the user to add external
JavaScript files to the WebPart that can run precisely before or after
the rendering of the template.

* Fixed the readme.md imaged broke in the previous commit

* Added the ExternalScripts.png picture to readme.md

* Updated readme.md

* Update the externalScripts image.

* Fixed an issue where sites/subsites were missing from the WebUrl dropdown

The search query responsible for feeding the Web Url dropdown had no
rowLimit specified, which was limiting the amount of returned urls to 10
since this is the default rowLimit. Updated the rowLimit to 500 so all
sites can be displayed in the toolpart!

* Update README.md