(#3299) fixed nesting of Headers
The NavLinkBuilder was modifying arrays in place,
as well as returning the modifications. While this was probably not
the cause of the error, I fixed this by not returning the modifications
anymore.

I changed the algorithm from counting the depth up, to counting the
order down to 0. Thus, I could remove the depth entirely.